requirements.txt
fastapi
pydantic
ultralytics
Pillow
python-multipart
math
datetime
pandas
scikit-learn
torch
numpy
matplotlib
seaborn
